+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `shownToUser` method to Companion Ads, which should be fired when the ad is shown to the user
 - `hiddenFromUser` method to Companion Ads, which should be fired when the ad is hidden from the user
 - `UNKNOWN_FORMAT` error (code `1011`)
+- `YospaceConfiguration.debugYospaceSdk` to enable debug logs of the Yospace SDK without enabling logs for the `BitmovinYospacePlayer` (helpful for Yospace validation)
+- Support for tracking muted/unmuted changes
 
 ## Changed
 
@@ -23,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `getYospaceManager` to `getYospaceSession`
 - Build chain to use webpack, eslint and prettier
 - Yospace SDK to be imported via npm
+- Events of the `BitmovinYospacePlayer` are now triggered asynchronously
 
 ## Removed